+++ lib/src/chat_l10n.dart
@@ -204,3 +204,47 @@ class ChatL10nUk extends ChatL10n {
           sendButtonAccessibilityLabel: sendButtonAccessibilityLabel,
         );
 }
+
+/// Simplified Chinese l10n which extends [ChatL10n]
+@immutable
+class ChatL10nZhCN extends ChatL10n {
+  /// Creates Simplified Chinese l10n. Use this constructor if you want to
+  /// override only a couple of properties, otherwise create a new class
+  /// which extends [ChatL10n]
+  const ChatL10nZhCN({
+    String attachmentButtonAccessibilityLabel = '发送媒体文件',
+    String emptyChatPlaceholder = '暂无消息',
+    String fileButtonAccessibilityLabel = '文件',
+    String inputPlaceholder = '输入消息',
+    String sendButtonAccessibilityLabel = '发送',
+  }) : super(
+          attachmentButtonAccessibilityLabel:
+              attachmentButtonAccessibilityLabel,
+          emptyChatPlaceholder: emptyChatPlaceholder,
+          fileButtonAccessibilityLabel: fileButtonAccessibilityLabel,
+          inputPlaceholder: inputPlaceholder,
+          sendButtonAccessibilityLabel: sendButtonAccessibilityLabel,
+        );
+}
+
+/// Traditional Chinese l10n which extends [ChatL10n]
+@immutable
+class ChatL10nZhTW extends ChatL10n {
+  /// Creates Traditional Chinese l10n. Use this constructor if you want to
+  /// override only a couple of properties, otherwise create a new class
+  /// which extends [ChatL10n]
+  const ChatL10nZhTW({
+    String attachmentButtonAccessibilityLabel = '傳送媒體',
+    String emptyChatPlaceholder = '還沒有訊息在這裡',
+    String fileButtonAccessibilityLabel = '檔案',
+    String inputPlaceholder = '輸入訊息',
+    String sendButtonAccessibilityLabel = '傳送',
+  }) : super(
+          attachmentButtonAccessibilityLabel:
+              attachmentButtonAccessibilityLabel,
+          emptyChatPlaceholder: emptyChatPlaceholder,
+          fileButtonAccessibilityLabel: fileButtonAccessibilityLabel,
+          inputPlaceholder: inputPlaceholder,
+          sendButtonAccessibilityLabel: sendButtonAccessibilityLabel,
+        );
+}