+++ lib/src/widgets/chat.dart
@@ -9,6 +9,7 @@ import '../chat_l10n.dart';
 import '../chat_theme.dart';
 import '../conditional/conditional.dart';
 import '../models/date_header.dart';
+import '../models/emoji_enlargement_behavior.dart';
 import '../models/message_spacer.dart';
 import '../models/preview_image.dart';
 import '../models/send_button_visibility_mode.dart';
@@ -36,23 +37,30 @@ class Chat extends StatefulWidget {
     this.dateHeaderThreshold = 900000,
     this.dateLocale,
     this.disableImageGallery,
+    this.emojiEnlargementBehavior = EmojiEnlargementBehavior.multi,
     this.emptyState,
     this.fileMessageBuilder,
     this.groupMessagesThreshold = 60000,
+    this.hideBackgroundOnEmojiMessages = true,
     this.imageMessageBuilder,
     this.isAttachmentUploading,
     this.isLastPage,
     this.l10n = const ChatL10nEn(),
     required this.messages,
     this.onAttachmentPressed,
+    this.onAvatarTap,
+    this.onBackgroundTap,
     this.onEndReached,
     this.onEndReachedThreshold,
     this.onMessageLongPress,
+    this.onMessageStatusLongPress,
+    this.onMessageStatusTap,
     this.onMessageTap,
     this.onPreviewDataFetched,
     required this.onSendPressed,
     this.onTextChanged,
     this.onTextFieldTap,
+    this.scrollPhysics,
     this.sendButtonVisibilityMode = SendButtonVisibilityMode.editing,
     this.showUserAvatars = false,
     this.showUserNames = false,
@@ -109,6 +117,9 @@ class Chat extends StatefulWidget {
   /// Disable automatic image preview on tap.
   final bool? disableImageGallery;
 
+  /// See [Message.emojiEnlargementBehavior]
+  final EmojiEnlargementBehavior emojiEnlargementBehavior;
+
   /// Allows you to change what the user sees when there are no messages.
   /// `emptyChatPlaceholder` and `emptyChatPlaceholderTextStyle` are ignored
   /// in this case.
@@ -122,6 +133,9 @@ class Chat extends StatefulWidget {
   /// Default value is 1 minute, 60000 ms. When time between two messages
   /// is lower than this threshold, they will be visually grouped.
   final int groupMessagesThreshold;
+
+  /// See [Message.hideBackgroundOnEmojiMessages]
+  final bool hideBackgroundOnEmojiMessages;
 
   /// See [Message.imageMessageBuilder]
   final Widget Function(types.ImageMessage, {required int messageWidth})?
@@ -144,6 +158,12 @@ class Chat extends StatefulWidget {
   /// See [Input.onAttachmentPressed]
   final void Function()? onAttachmentPressed;
 
+  /// See [Message.onAvatarTap]
+  final void Function(types.User)? onAvatarTap;
+
+  /// Called when user taps on background
+  final void Function()? onBackgroundTap;
+
   /// See [ChatList.onEndReached]
   final Future<void> Function()? onEndReached;
 
@@ -152,6 +172,12 @@ class Chat extends StatefulWidget {
 
   /// See [Message.onMessageLongPress]
   final void Function(types.Message)? onMessageLongPress;
+
+  /// See [Message.onMessageStatusLongPress]
+  final void Function(types.Message)? onMessageStatusLongPress;
+
+  /// See [Message.onMessageStatusTap]
+  final void Function(types.Message)? onMessageStatusTap;
 
   /// See [Message.onMessageTap]
   final void Function(types.Message)? onMessageTap;
@@ -168,6 +194,9 @@ class Chat extends StatefulWidget {
 
   /// See [Input.onTextFieldTap]
   final void Function()? onTextFieldTap;
+
+  /// See [ChatList.scrollPhysics]
+  final ScrollPhysics? scrollPhysics;
 
   /// See [Input.sendButtonVisibilityMode]
   final SendButtonVisibilityMode sendButtonVisibilityMode;
@@ -314,10 +343,7 @@ class _ChatState extends State<Chat> {
     if (object is DateHeader) {
       return Container(
         alignment: Alignment.center,
-        margin: const EdgeInsets.only(
-          bottom: 32,
-          top: 16,
-        ),
+        margin: widget.theme.dateDividerMargin,
         child: Text(
           object.text,
           style: widget.theme.dateDividerTextStyle,
@@ -339,11 +365,16 @@ class _ChatState extends State<Chat> {
         key: ValueKey(message.id),
         bubbleBuilder: widget.bubbleBuilder,
         customMessageBuilder: widget.customMessageBuilder,
+        emojiEnlargementBehavior: widget.emojiEnlargementBehavior,
         fileMessageBuilder: widget.fileMessageBuilder,
+        hideBackgroundOnEmojiMessages: widget.hideBackgroundOnEmojiMessages,
         imageMessageBuilder: widget.imageMessageBuilder,
         message: message,
         messageWidth: _messageWidth,
+        onAvatarTap: widget.onAvatarTap,
         onMessageLongPress: widget.onMessageLongPress,
+        onMessageStatusLongPress: widget.onMessageStatusLongPress,
+        onMessageStatusTap: widget.onMessageStatusTap,
         onMessageTap: (tappedMessage) {
           if (tappedMessage is types.ImageMessage &&
               widget.disableImageGallery != true) {
@@ -412,8 +443,10 @@ class _ChatState extends State<Chat> {
                               child: _emptyStateBuilder(),
                             )
                           : GestureDetector(
-                              onTap: () =>
-                                  FocusManager.instance.primaryFocus?.unfocus(),
+                              onTap: () {
+                                FocusManager.instance.primaryFocus?.unfocus();
+                                widget.onBackgroundTap?.call();
+                              },
                               child: LayoutBuilder(
                                 builder: (BuildContext context,
                                         BoxConstraints constraints) =>
@@ -425,6 +458,7 @@ class _ChatState extends State<Chat> {
                                   onEndReached: widget.onEndReached,
                                   onEndReachedThreshold:
                                       widget.onEndReachedThreshold,
+                                  scrollPhysics: widget.scrollPhysics,
                                 ),
                               ),
                             ),