+++ lib/src/widgets/message.dart
@@ -1,5 +1,6 @@
 import 'package:flutter/material.dart';
 import 'package:flutter_chat_types/flutter_chat_types.dart' as types;
+import '../models/emoji_enlargement_behavior.dart';
 import '../util.dart';
 import 'file_message.dart';
 import 'image_message.dart';
@@ -16,11 +17,16 @@ class Message extends StatelessWidget {
     Key? key,
     this.bubbleBuilder,
     this.customMessageBuilder,
+    required this.emojiEnlargementBehavior,
     this.fileMessageBuilder,
+    required this.hideBackgroundOnEmojiMessages,
     this.imageMessageBuilder,
     required this.message,
     required this.messageWidth,
+    this.onAvatarTap,
     this.onMessageLongPress,
+    this.onMessageStatusLongPress,
+    this.onMessageStatusTap,
     this.onMessageTap,
     this.onPreviewDataFetched,
     required this.roundBorder,
@@ -47,9 +53,17 @@ class Message extends StatelessWidget {
   final Widget Function(types.CustomMessage, {required int messageWidth})?
       customMessageBuilder;
 
+  /// Controls the enlargement behavior of the emojis in the
+  /// [types.TextMessage].
+  /// Defaults to [EmojiEnlargementBehavior.multi].
+  final EmojiEnlargementBehavior emojiEnlargementBehavior;
+
   /// Build a file message inside predefined bubble
   final Widget Function(types.FileMessage, {required int messageWidth})?
       fileMessageBuilder;
+
+  /// Hide background for messages containing only emojis.
+  final bool hideBackgroundOnEmojiMessages;
 
   /// Build an image message inside predefined bubble
   final Widget Function(types.ImageMessage, {required int messageWidth})?
@@ -61,8 +75,17 @@ class Message extends StatelessWidget {
   /// Maximum message width
   final int messageWidth;
 
+  // Called when uses taps on an avatar
+  final void Function(types.User)? onAvatarTap;
+
   /// Called when user makes a long press on any message
   final void Function(types.Message)? onMessageLongPress;
+
+  /// Called when user makes a long press on status icon in any message
+  final void Function(types.Message)? onMessageStatusLongPress;
+
+  /// Called when user taps on status icon in any message
+  final void Function(types.Message)? onMessageStatusTap;
 
   /// Called when user taps on any message
   final void Function(types.Message)? onMessageTap;
@@ -107,23 +130,26 @@ class Message extends StatelessWidget {
     return showAvatar
         ? Container(
             margin: const EdgeInsets.only(right: 8),
-            child: CircleAvatar(
-              backgroundColor: hasImage
-                  ? InheritedChatTheme.of(context)
-                      .theme
-                      .userAvatarImageBackgroundColor
-                  : color,
-              backgroundImage:
-                  hasImage ? NetworkImage(message.author.imageUrl!) : null,
-              radius: 16,
-              child: !hasImage
-                  ? Text(
-                      initials,
-                      style: InheritedChatTheme.of(context)
-                          .theme
-                          .userAvatarTextStyle,
-                    )
-                  : null,
+            child: GestureDetector(
+              onTap: () => onAvatarTap?.call(message.author),
+              child: CircleAvatar(
+                backgroundColor: hasImage
+                    ? InheritedChatTheme.of(context)
+                        .theme
+                        .userAvatarImageBackgroundColor
+                    : color,
+                backgroundImage:
+                    hasImage ? NetworkImage(message.author.imageUrl!) : null,
+                radius: 16,
+                child: !hasImage
+                    ? Text(
+                        initials,
+                        style: InheritedChatTheme.of(context)
+                            .theme
+                            .userAvatarTextStyle,
+                      )
+                    : null,
+              ),
             ),
           )
         : const SizedBox(width: 40);
@@ -133,6 +159,7 @@ class Message extends StatelessWidget {
     BuildContext context,
     BorderRadius borderRadius,
     bool currentUserIsAuthor,
+    bool enlargeEmojis,
   ) {
     return bubbleBuilder != null
         ? bubbleBuilder!(
@@ -140,19 +167,21 @@ class Message extends StatelessWidget {
             message: message,
             nextMessageInGroup: roundBorder,
           )
-        : Container(
-            decoration: BoxDecoration(
-              borderRadius: borderRadius,
-              color: !currentUserIsAuthor ||
-                      message.type == types.MessageType.image
-                  ? InheritedChatTheme.of(context).theme.secondaryColor
-                  : InheritedChatTheme.of(context).theme.primaryColor,
-            ),
-            child: ClipRRect(
-              borderRadius: borderRadius,
-              child: _messageBuilder(),
-            ),
-          );
+        : enlargeEmojis && hideBackgroundOnEmojiMessages
+            ? _messageBuilder()
+            : Container(
+                decoration: BoxDecoration(
+                  borderRadius: borderRadius,
+                  color: !currentUserIsAuthor ||
+                          message.type == types.MessageType.image
+                      ? InheritedChatTheme.of(context).theme.secondaryColor
+                      : InheritedChatTheme.of(context).theme.primaryColor,
+                ),
+                child: ClipRRect(
+                  borderRadius: borderRadius,
+                  child: _messageBuilder(),
+                ),
+              );
   }
 
   Widget _messageBuilder() {
@@ -181,6 +210,8 @@ class Message extends StatelessWidget {
                 showName: showName,
               )
             : TextMessage(
+                emojiEnlargementBehavior: emojiEnlargementBehavior,
+                hideBackgroundOnEmojiMessages: hideBackgroundOnEmojiMessages,
                 message: textMessage,
                 onPreviewDataFetched: onPreviewDataFetched,
                 showName: showName,
@@ -243,6 +274,11 @@ class Message extends StatelessWidget {
   Widget build(BuildContext context) {
     final _user = InheritedUser.of(context).user;
     final _currentUserIsAuthor = _user.id == message.author.id;
+    final _enlargeEmojis =
+        emojiEnlargementBehavior != EmojiEnlargementBehavior.never &&
+            message is types.TextMessage &&
+            isConsistsOfEmojis(
+                emojiEnlargementBehavior, message as types.TextMessage);
     final _messageBorderRadius =
         InheritedChatTheme.of(context).theme.messageBorderRadius;
     final _borderRadius = BorderRadius.only(
@@ -284,6 +320,7 @@ class Message extends StatelessWidget {
                     context,
                     _borderRadius,
                     _currentUserIsAuthor,
+                    _enlargeEmojis,
                   ),
                 ),
               ],
@@ -291,14 +328,15 @@ class Message extends StatelessWidget {
           ),
           if (_currentUserIsAuthor)
             Padding(
-              padding: const EdgeInsets.symmetric(horizontal: 4),
-              child: Center(
-                child: SizedBox(
-                  height: 16,
-                  width: 16,
-                  child: showStatus ? _statusBuilder(context) : null,
-                ),
-              ),
+              padding: InheritedChatTheme.of(context).theme.statusIconPadding,
+              child: showStatus
+                  ? GestureDetector(
+                      onLongPress: () =>
+                          onMessageStatusLongPress?.call(message),
+                      onTap: () => onMessageStatusTap?.call(message),
+                      child: _statusBuilder(context),
+                    )
+                  : null,
             ),
         ],
       ),