+++ test/flutter_chat_ui_test.dart
@@ -1,7 +1,13 @@
+import 'file_message.test.dart' as file_message_test;
+import 'image_message.test.dart' as image_message_test;
 import 'input.test.dart' as input_test;
+import 'text_message.test.dart' as text_message_test;
 import 'util.test.dart' as util_test;
 
 void main() {
+  file_message_test.main();
+  image_message_test.main();
   input_test.main();
+  text_message_test.main();
   util_test.main();
 }